+++ CHANGES.rst
@@ -4,6 +4,10 @@ Version History
 Version: 1.2.7 (unreleased)
 -------------------------------------------------------------------------------
 
+BACKWARD-INCOMPATIBLE:
+
+* Replace old-style tag-expressions with `cucumber-tag-expressions`_
+
 GOALS:
 
 - Improve support for Windows (continued)
@@ -23,6 +27,7 @@ DOCUMENTATION:
 
 ENHANCEMENTS:
 
+* Use `cucumber-tag-expressions`_ with tag-matching extension (superceeds: old-style tag-expressions)
 * issue #678: Scenario Outline: Support tags with commas and semicolons (provided by: lawnmowerlatte, pull #679)
 * issue #675: Feature files cannot be found within symlink directories (provided by: smadness, pull #680)
 
@@ -39,6 +44,9 @@ MINOR:
 * pull  #655: Use pytest instead of py.test per upstream recommendation (provided by: scop)
 * issue #654: tox.ini: pypi.python.org -> pypi.org (submitted by: pradyunsg)
 * pull  #660: Fix minor typos (provided by: rrueth)
+
+
+.. _`cucumber-tag-expressions`: https://pypi.org/project/cucumber-tag-expressions/
 
 
 Version: 1.2.6 (2018-02-25)